feat(create-starknet): added command arguments support
