+++ packages/create-starknet/nodemon.json
@@ -1,6 +1,6 @@
 {
-  "watch": ["src"],
+  "watch": ["./"],
   "ext": ".ts,.js",
   "ignore": [],
-  "exec": "npx ts-node ./src/index.ts"
+  "exec": "npx ts-node ./index.ts"
 }
